requirements.txt
streamlit
requests
beautifulsoup4
crewai==0.1.32
pandas==1.5.1
langchain-groq
python-dotenv
pdfplumber
matplotlib
seaborn